+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - add Bento rule
 - #179 add ezyVet rule (thanks @oneofthedamons)
+- #180 add Emarsys (thanks @oneofthedamons)
 
 ## [0.7.0] - 2022-07-02